+++ cmd/ethrex/Cargo.toml
@@ -30,7 +30,6 @@ local-ip-address = "0.6"
 tokio-util.workspace = true
 redb = { workspace = true, optional = true }
 lazy_static.workspace = true
-secp256k1.workspace = true
 
 cfg-if = "1.0.0"
 
@@ -49,4 +48,4 @@ metrics = ["ethrex-blockchain/metrics", "ethrex-l2/metrics"]
 libmdbx = ["ethrex-storage/libmdbx"]
 redb = ["dep:redb", "ethrex-storage/redb"]
 blst = ["ethrex-vm/blst"]
-l2 = ["dep:ethrex-l2", "ethrex-vm/l2"]
+l2 = ["dep:ethrex-l2", "ethrex-vm/l2", "dep:ethrex-dev"]
