+++ crates/blockchain/dev/Cargo.toml
@@ -7,6 +7,8 @@ edition = "2021"
 
 [dependencies]
 ethrex-rpc.workspace = true
+ethrex-common.workspace = true
+ethrex-sdk.workspace = true
 
 tokio.workspace = true
 tracing.workspace = true
@@ -21,6 +23,7 @@ reqwest = { version = "0.12.7", features = ["json"] }
 envy = "0.4.2"
 keccak-hash.workspace = true
 sha2 = "0.10.8"
+secp256k1.workspace = true
 
 [lib]
 path = "./dev.rs"
